+++ CHANGELOG.md
@@ -17,8 +17,14 @@
 * Formatting
 * Begin factoring some of the code into separate files - this is a work in progress
 
-## 3.0 -- 2022-09-14
+## 2.0.0.0 -- 2022-09-14
 
 * Major refactoring
 * Expanded API
 * Generally tidier codebase
+
+## 2.1.0.0 -- 2022-09-16
+
+* Rework parts of the API to use the record datatypes
+* Move all of internal into just Database/HsPOS
+* Add Util module for general utility functions